Cleanup (2) / Filters bugged in guidance[backtrack purposes
