+++ resources/css/guidance_student-profile.css
@@ -1,0 +1,24 @@
+/* Search and filter styling */
+.search-filter-section {
+    background: white;
+    border-radius: 0.5rem;
+    padding: 1.5rem;
+    margin-bottom: 2rem;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+  }
+  
+  /* Student card styling */
+  .student-card {
+    transition: transform 0.2s ease;
+  }
+  
+  .student-card:hover {
+    transform: translateY(-2px);
+    box-shadow: 0 4px 8px rgba(0,0,0,0.15);
+  }
+  
+  /* Camera modal styling */
+  video {
+    border: 2px solid var(--secondary-color);
+    border-radius: 0.5rem;
+  }
