+++ vite.config.js
@@ -18,6 +18,10 @@ export default defineConfig({
                 'resources/css/manage_users.css',
                 'resources/css/index_student.css',
                 'resources/css/index_guidance.css',
+                'resources/js/guidance_student-profile.js',
+                'resources/css/guidance_student-profile.css',
+                'resources/js/guidance_student-violations.js',
+                'resources/css/guidance_student-violations.css',
                 'resources/css/student_violations.css',
                 'resources/js/landingpage.js',
                 'resources/js/enroll.js',
